Support a URL scheme without port
In some cases, such as connecting the plugin to a cloud-hosted RabbitMQ at
lshift.net, the API URL should be constructed without the port in order for
connection to be successful.

Before the fix, the URLs looks like this:
``http://bigwig.lshift.net/management/12345:80/api/channels``

After the fix, like this:
``http://bigwig.lshift.net/management/12345/api/channels``

Notice the ``:80`` that is added in the path part of the URL. To skip adding
the port number to the URL, the user should add the following to the config:
``port: none``.